test organizations internals

+++ support/src/testing.rs
@@ -35,7 +35,7 @@ parameter_types! {
 }
 
 pub mod primitives {
-    pub type AccountId = u64;
+    pub type AccountId = u128;
     pub type Balance = u64;
     pub type CurrencyId = u8;
 }
